nix_python/schemas.py
from pydantic import BaseModel, Extra


class Ping(BaseModel):
    ping: str


class ItemBase(BaseModel):
    title: str
    description: str | None = None


class ItemCreate(ItemBase):
    pass

    class Config:
        extra = Extra.forbid


class Item(ItemBase):
    id: int
    owner_id: int

    class Config:
        orm_mode = True


class UserBase(BaseModel):
    email: str


class UserCreate(UserBase):
    password: str

    class Config:
        extra = Extra.forbid


class User(UserBase):
    id: int
    is_active: bool
    items: list[Item] = []

    class Config:
        orm_mode = True
