pyproject.toml
[tool.poetry]
name = "nix-python"
version = "0.1.0"
description = ""
authors = ["werner mendizabal <nonameentername@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.4"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
psycopg2 = "^2.9.5"
alembic = "^1.10.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "nix_python.main:start"
